Fixing build issues
