+++ ipc/Cargo.toml
@@ -15,9 +15,11 @@ slab   = "0.1.0"
 bytes  = "0.3.0"
 miow = "0.1"
 log = "0.3"
-env_logger = "0.3"
-lazy_static = "0.2"
 rand = "0.3"
+
+[dev-dependencies]
+lazy_static = "0.2"
+env_logger = "0.4"
 
 [features]
 default = ["jsonrpc-core/reactor"]
