+++ ipc/src/tests.rs
@@ -16,8 +16,31 @@
 
 use jsonrpc_core::IoHandler;
 use super::Server;
-use std;
+use std::{self, env};
 use rand::{thread_rng, Rng};
+use log::LogLevelFilter;
+use env_logger::LogBuilder;
+
+lazy_static! {
+	static ref LOG_DUMMY: bool = {
+		let mut builder = LogBuilder::new();
+		builder.filter(None, LogLevelFilter::Info);
+
+		if let Ok(log) = env::var("RUST_LOG") {
+			builder.parse(&log);
+		}
+
+		if let Ok(_) = builder.init() {
+			println!("logger initialized");
+		}
+		true
+	};
+}
+
+/// Intialize log with default settings
+pub fn init_log() {
+	let _ = *LOG_DUMMY;
+}
 
 #[cfg(test)]
 pub fn dummy_io_handler() -> IoHandler {
@@ -95,7 +118,7 @@ pub fn test_reqrep() {
 
 #[test]
 pub fn test_reqrep_two_sequental_connections() {
-	super::init_log();
+	init_log();
 
 	let addr = random_ipc_endpoint();
 	let io = dummy_io_handler();
@@ -137,7 +160,7 @@ pub fn test_reqrep_three_sequental_connections() {
 #[test]
 #[ignore]
 pub fn test_reqrep_100_connections() {
-	super::init_log();
+	init_log();
 
 	let addr = random_ipc_endpoint();
 	let io = dummy_io_handler();
@@ -155,7 +178,7 @@ pub fn test_reqrep_100_connections() {
 #[test]
 #[ignore]
 pub fn test_reqrep_10_pubsub() {
-	super::init_log();
+	init_log();
 
 	let addr = random_ipc_endpoint();
 	let io = dummy_io_handler();
